[DSIP-97][API/WORK] datasource driver version
